+++ test/api-client.js
@@ -68,18 +68,25 @@ exports.signupAs = function signupAs(user, callback) {
 };
 // HTTP request wrappers
 
-exports.get = function (jar, url, callback) {
+exports.getRaw = function (jar, url, callback) {
   request.get(
     { jar, url: `${URL_PREFIX}${url}` },
     function (error, response, body) {
-      try {
-        body = JSON.parse(body);
-      } catch (e) {
-        console.error(e);
-      }
-      callback(error, { response, body, jar });
+      callback(error, { response, body });
     }
   );
+};
+
+exports.get = function (jar, url, callback) {
+  exports.getRaw(jar, url, function (error, { response, body }) {
+    let parsedBody = undefined;
+    try {
+      parsedBody = JSON.parse(body);
+    } catch (e) {
+      error = error | e;
+    }
+    callback(error, { response, body: parsedBody, jar });
+  });
 };
 
 // USER